fixed leak in RADEC
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@4830 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/utils/rts2valuestat.h
@@ -33,7 +33,7 @@
 class Rts2ValueDoubleStat:public Rts2ValueDouble
 {
 	private:
-		size_t numMes;
+		int numMes;
 		double mode;
 		double min;
 		double max;
